refactor(lib): break the client/users import cycle via lib/errors.py
CodeQL flags the cycle client.py → users.py → (lazy) client.py. The
typed transport errors (CaptchaError, AuthenticationError) and
_sanitize_error move to a new lib/errors.py that both sides import;
client.py re-exports every name, so all existing
`from lib.client import ...` call sites keep working. users.py now has
no client import at all, lazy or otherwise.
