+++ skills/jira-communication/scripts/lib/client.py
@@ -9,6 +9,7 @@ from requests.adapters import HTTPAdapter
 from urllib3.util.retry import Retry
 
 from .config import get_auth_mode, is_cloud_url, load_config, validate_config
+from .errors import AuthenticationError, CaptchaError, _sanitize_error  # noqa: F401  (re-exported)
 from .users import find_users, is_cloud_client
 
 # Default timeout for all Jira API requests (seconds)
@@ -242,18 +243,6 @@ def resolve_subtask_type(client, project_key: str, requested_type: str) -> str |
 # === INLINE_START: client ===
 
 
-class CaptchaError(Exception):
-    """Error raised when Jira requires CAPTCHA resolution.
-
-    This happens when Jira Server/DC detects suspicious login activity
-    and requires the user to complete a CAPTCHA challenge in the web UI.
-    """
-
-    def __init__(self, message: str, login_url: str):
-        super().__init__(message)
-        self.login_url = login_url
-
-
 class SessionExpiredError(Exception):
     """Raised when a 200 OK response carries an HTML session-expiry or login page.
 
@@ -261,14 +250,6 @@ class SessionExpiredError(Exception):
     200 OK with Content-Type: text/html and no Content-Disposition: attachment.
     Without this check the HTML body would be silently written to disk as if it
     were the requested file.
-    """
-
-
-class AuthenticationError(Exception):
-    """Raised when Jira returns 401 or 403 on an authenticated request.
-
-    Provides a typed alternative to inspecting raw HTTP status codes or
-    string-matching error messages for authentication failures.
     """
 
 
@@ -695,29 +676,6 @@ def get_jira_client(
                 f"    - JIRA_USERNAME is your email (Cloud) or username (Server/DC)\n"
                 f"    - JIRA_API_TOKEN is valid\n"
             ) from e
-
-
-def _sanitize_error(message: str) -> str:
-    """Remove potential credential fragments from error messages.
-
-    Uses regex to redact values after sensitive keys, rather than a simple
-    denylist check that discards the entire message.
-    """
-    # Redact values following sensitive keys (e.g., "token=abc123" → "token=***")
-    # First handle "Authorization: <scheme> <token>" as a single unit
-    sanitized = re.sub(
-        r"(authorization:\s*)\S+(?:\s+\S+)?",
-        r"\1***",
-        message,
-        flags=re.IGNORECASE,
-    )
-    sanitized = re.sub(
-        r"(bearer |basic |token=|password=|api_token=|api_key=|secret=|access_token=|private_token=|apikey=|auth_token=)\S+",
-        r"\1***",
-        sanitized,
-        flags=re.IGNORECASE,
-    )
-    return sanitized
 
 
 # === INLINE_END: client ===
